third_party/blink/web_tests/platform/generic/external/wpt/content-security-policy/securitypolicyviolation/blockeduri-eval-expected.txt

This is a testharness.js-based test.
FAIL Eval violations have a blockedURI of 'eval' assert_equals: expected 12 but got 13
Harness: the test ran to completion.
